Polishing - refactor code for reuse.
